Options.Bind dont work with complex objects, so, "ServiceOptions" is pure now
